+++ Cargo.toml
@@ -22,11 +22,10 @@ hdf5 = "0.8.1"
 lazy_static = "1.4.0"
 mpi = { version = "0.6", default-features = false, features = ["derive"] }
 ndarray = "0.15.6"
-once_cell = "1.13.0"
-paste = "1.0.8"
 serde = {version = "1.0.143", features = ["derive"] }
 serde_yaml = "0.9.9"
 derive_named = { path = "derive_named" }
+diman = { path = "diman", features = ["mpi", "hdf5", "serde"] }
 rand = "0.8.5"
 
 [dev-dependencies]
