+++ diman/src/mpi.rs
