+++ diman/src/sampling.rs
