+++ src/io/input/tests.rs
@@ -14,9 +14,9 @@ use crate::io::to_dataset::ToDataset;
 use crate::prelude::Mass;
 use crate::prelude::WorldRank;
 use crate::prelude::WorldSize;
+use crate::test_utils::assert_is_close;
 use crate::test_utils::run_system_on_world;
 use crate::test_utils::tests_path;
-use crate::units::assert_is_close;
 use crate::units::{self};
 
 #[test]
